9668 removed all the additional logs and change token caching calculation AB#9668 AB#9671

+++ UKHO.ExchangeSetService.API/UKHO.ExchangeSetService.Common/Configuration/EssManagedIdentityConfiguration.cs
@@ -6,6 +6,6 @@ namespace UKHO.ExchangeSetService.Common.Configuration
     public class EssManagedIdentityConfiguration
     {
         public string ClientId { get; set; }
-        public int DeductTokenExpiryMinutes { get; set; }
+        public double DeductTokenExpiryMinutes { get; set; }
     }
 }
